Removed ShareAppRepositoryImpl, ContactSupportRepositoryImpl and LicenseAgreementRepositoryImpl. All logic moved to SettingsViemModel
