+++ app/src/main/java/com/example/yp_playlist_maker/creator/Creator.kt
@@ -2,35 +2,23 @@ package com.example.yp_playlist_maker.creator
 
 import android.content.Context
 import android.net.ConnectivityManager
-import com.example.yp_playlist_maker.settings.data.impl.AppThemeImpl
 import com.example.yp_playlist_maker.player.data.impl.PlayTrackRepositoryImpl
+import com.example.yp_playlist_maker.player.domain.api.PlayTrackInteractor
+import com.example.yp_playlist_maker.player.domain.api.PlayTrackRepository
+import com.example.yp_playlist_maker.player.domain.impl.PlayTrackInteractorImpl
 import com.example.yp_playlist_maker.search.data.impl.SearchHistoryRepositoryImpl
 import com.example.yp_playlist_maker.search.data.impl.TrackRepositoryImpl
 import com.example.yp_playlist_maker.search.data.network.RetrofitNetworkClient
-import com.example.yp_playlist_maker.settings.domain.api.AppThemeInteractor
-import com.example.yp_playlist_maker.player.domain.api.PlayTrackInteractor
 import com.example.yp_playlist_maker.search.domain.api.SearchHistoryInteractor
-import com.example.yp_playlist_maker.search.domain.api.TrackInteractor
-import com.example.yp_playlist_maker.settings.domain.api.AppThemeRepository
-import com.example.yp_playlist_maker.player.domain.api.PlayTrackRepository
 import com.example.yp_playlist_maker.search.domain.api.SearchHistoryRepository
+import com.example.yp_playlist_maker.search.domain.api.TrackInteractor
 import com.example.yp_playlist_maker.search.domain.api.TrackRepository
-import com.example.yp_playlist_maker.settings.domain.impl.AppThemeInteractorImpl
-import com.example.yp_playlist_maker.player.domain.impl.PlayTrackInteractorImpl
 import com.example.yp_playlist_maker.search.domain.impl.SearchHistoryInteractorImpl
 import com.example.yp_playlist_maker.search.domain.impl.TrackInteractorImpl
-import com.example.yp_playlist_maker.sharing.data.impl.ContactSupportRepositoryImpl
-import com.example.yp_playlist_maker.sharing.data.impl.LicenseAgreementRepositoryImpl
-import com.example.yp_playlist_maker.sharing.data.impl.ShareAppRepositoryImpl
-import com.example.yp_playlist_maker.sharing.domain.api.interactor.ContactSupportInteractor
-import com.example.yp_playlist_maker.sharing.domain.api.interactor.LicenseAgreementInteractor
-import com.example.yp_playlist_maker.sharing.domain.api.interactor.ShareAppInteractor
-import com.example.yp_playlist_maker.sharing.domain.api.repository.ContactSupportRepository
-import com.example.yp_playlist_maker.sharing.domain.api.repository.LicenseAgreementRepository
-import com.example.yp_playlist_maker.sharing.domain.api.repository.ShareAppRepository
-import com.example.yp_playlist_maker.sharing.domain.impl.ContactSupportInteractorImpl
-import com.example.yp_playlist_maker.sharing.domain.impl.LicenseAgreementInteractorImpl
-import com.example.yp_playlist_maker.sharing.domain.impl.ShareAppInteractorImpl
+import com.example.yp_playlist_maker.settings.data.impl.AppThemeImpl
+import com.example.yp_playlist_maker.settings.domain.api.AppThemeInteractor
+import com.example.yp_playlist_maker.settings.domain.api.AppThemeRepository
+import com.example.yp_playlist_maker.settings.domain.impl.AppThemeInteractorImpl
 
 object Creator {
 
@@ -76,33 +64,6 @@ object Creator {
 
     fun providePlayTrackInteractor() : PlayTrackInteractor {
         return PlayTrackInteractorImpl(getPlayTrackRepository())
-    }
-
-    // Поделиться приложением
-    private fun getShareAppRepository() : ShareAppRepository {
-        return ShareAppRepositoryImpl(appContext)
-    }
-
-    fun provideShareAppInteractor() : ShareAppInteractor {
-        return ShareAppInteractorImpl(getShareAppRepository())
-    }
-
-    // Обратиться в поддержку
-    private fun getContactSupportRepository() : ContactSupportRepository {
-        return ContactSupportRepositoryImpl(appContext)
-    }
-
-    fun provideContactSupportInteractor() : ContactSupportInteractor {
-        return ContactSupportInteractorImpl(getContactSupportRepository())
-    }
-
-    // Пользовательское соглашение
-    private fun getLicenseAgreementRepository(): LicenseAgreementRepository {
-        return LicenseAgreementRepositoryImpl(appContext)
-    }
-
-    fun provideLicenseAgreementInteractor(): LicenseAgreementInteractor {
-        return LicenseAgreementInteractorImpl(getLicenseAgreementRepository())
     }
 
 }
