+++ app/src/main/java/com/example/yp_playlist_maker/settings/ui/SettingsActivity.kt
@@ -35,17 +35,17 @@ class SettingsActivity : AppCompatActivity() {
 
         // Поделиться приложением
         binding.flShare.setOnClickListener {
-            viewModel.share()
+            viewModel.share(this)
         }
 
         // Написать в поддержку
         binding.flContactSupport.setOnClickListener {
-            viewModel.contact()
+            viewModel.contact(this)
         }
 
         // Лицензионное соглашение
         binding.flLicenseAgreement.setOnClickListener {
-            viewModel.getLicenseAgreement()
+            viewModel.getLicenseAgreement(this)
         }
 
     }
