--- app/src/main/java/com/example/yp_playlist_maker/sharing/domain/api/interactor/ContactSupportInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.yp_playlist_maker.sharing.domain.api.interactor
-
-interface ContactSupportInteractor {
-
-    fun contact()
-
-}
