--- app/src/main/java/com/example/yp_playlist_maker/sharing/domain/api/interactor/LicenseAgreementInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.yp_playlist_maker.sharing.domain.api.interactor
-
-interface LicenseAgreementInteractor {
-
-    fun getLicenseAgreement()
-
-}
