--- app/src/main/java/com/example/yp_playlist_maker/sharing/domain/api/repository/LicenseAgreementRepository.kt
@@ -1,7 +1,0 @@
-package com.example.yp_playlist_maker.sharing.domain.api.repository
-
-interface LicenseAgreementRepository {
-
-    fun getLicenseAgreement()
-
-}
