--- app/src/main/java/com/example/yp_playlist_maker/sharing/domain/impl/ShareAppInteractorImpl.kt
@@ -1,12 +1,0 @@
-package com.example.yp_playlist_maker.sharing.domain.impl
-
-import com.example.yp_playlist_maker.sharing.domain.api.interactor.ShareAppInteractor
-import com.example.yp_playlist_maker.sharing.domain.api.repository.ShareAppRepository
-
-class ShareAppInteractorImpl(private val repository: ShareAppRepository) : ShareAppInteractor {
-
-    override fun share() {
-        repository.share()
-    }
-
-}
